# data/openapi.yaml
openapi: 3.0.0
info:
  title: 'php samle api'
  version: 1.0.0
servers:
  -
    url: '{schema}://localhost'
    description: 'OpenApi parameters'
    variables:
      schema:
        enum:
          - https
          - http
        default: https
paths:
  /api/home:
    get:
      description: 'find tag by search condition'
      operationId: /api/home/index
      responses:
        '200':
          description: 'An example resource'
  /api/tag:
    get:
      tags:
        - tag
      summary: 'find tag by search condition'
      description: 'Returns array tag'
      operationId: getTagsBySearchConditions
      parameters:
        -
          name: name
          in: query
          description: 'search by tag name'
          required: false
          schema:
            type: string
        -
          name: site_id
          in: query
          description: 'search by site id'
          required: false
          schema:
            type: integer
            format: int32
        -
          name: limit
          in: query
          description: 'maximum number of results to return'
          required: false
          schema:
            type: integer
            format: int32
        -
          name: offset
          in: query
          description: 'offset of results to return'
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: 'successful operation'
        '400':
          description: 'Invalid Search Conditions supplied'
  /api/tag_site:
    post:
      tags:
        - tag
      summary: 'create for tag site relations'
      description: 'add data in tag_site table '
      operationId: tag_site_create
      responses:
        '200':
          description: 'successful operation'
components:
  schemas:
    Site:
      title: 'site model'
      description: 'Site model'
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: site1
        url:
          type: string
          example: 'http://google.com'
      type: object
    Tag:
      title: 'Tag model'
      description: 'Tag model'
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: tag1
      type: object